fix(library): home card label goes through homeMountPath
Review finding: the card showed the raw slug while provisioning could
fall back to mounting at org/home for reserved slugs (output/upload/
public/home — all permitted by slug rules), leaving e.g. two cards
both named "public" that open different places. homeMountPath moved to
a dependency-free module (file-storage/home-mount.ts) imported by both
provisioning and the web bundle, so label and mount can't drift.
