+++ apps/mesh/src/file-storage/mount/provisioning.ts
@@ -23,6 +23,7 @@
  *     folder appear in the sandbox with no copy step.
  */
 
+import { homeMountPath } from "../home-mount";
 import { getPublicSets, publicVolumeForSet } from "../public-sets";
 
 /** Shape the daemon parses from ORGFS_CONFIG (mirrors OrgFsMountConfig). */
@@ -43,22 +44,6 @@ const DEFAULT_MOUNTS: ReadonlyArray<{ volume: string; path: string }> = [
   { volume: "outputs", path: ORG_FS_OUTPUTS_MOUNT_PATH },
   { volume: "uploads", path: ORG_FS_UPLOADS_MOUNT_PATH },
 ];
-
-/** Other names that live directly under `org/` — a slug matching one of
- *  these would shadow it. */
-const RESERVED_HOME_PATHS = new Set(["output", "upload", "public", "home"]);
-/** One safe path segment, no traversal/hidden dirs (mirrors thread-links). */
-const SAFE_MOUNT_SEGMENT = /^[A-Za-z0-9][A-Za-z0-9._-]*$/;
-
-/** Mount path for the org's home volume: the org's own slug (immutable),
- *  so the agent sees `org/<slug>/`. Falls back to a literal `home` when the
- *  slug would collide with another `org/` entry or isn't a safe segment. */
-export function homeMountPath(orgSlug: string): string {
-  if (RESERVED_HOME_PATHS.has(orgSlug) || !SAFE_MOUNT_SEGMENT.test(orgSlug)) {
-    return "home";
-  }
-  return orgSlug;
-}
 
 export function buildOrgFsConfig(opts: {
   baseUrl: string;
